+++ indispensable-josef/src/jvmTest/kotlin/at/asitplus/signum/indispensable/josef/JweEncryptedTest.kt
@@ -92,7 +92,7 @@ val JweEncryptedTest by testSuite {
     "JWE symmetric encryption" - {
         withData(JweAlgorithm.Symmetric.entries) { alg ->
             val plain = Random.nextBytes(32)
-            val key = alg.randomKey().toJsonWebKey().getOrThrow()
+            val key = alg.randomKey(randomnessSourceOverride = Random.Default).toJsonWebKey().getOrThrow()
             val ciphertext = key.encrypt(plain).getOrThrow()
             ciphertext.decrypt(key) shouldSucceedWith plain
         }