+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/symmetric/KeyGen.kt
@@ -122,5 +122,5 @@ fun <A : AuthCapability<KeyType.WithDedicatedMacKey>, I : NonceTrait> SymmetricE
         (this as SymmetricEncryptionAlgorithm<*, *, *>).keyFromInternal(encryptionKey, macKey)
     } as KmmResult<SymmetricKey<A, I, KeyType.WithDedicatedMacKey>>
 
-suspend fun SpecializedSymmetricEncryptionAlgorithm.randomKey() =
-    algorithm.randomKey()
+suspend fun SpecializedSymmetricEncryptionAlgorithm.randomKey(randomnessSourceOverride : Random? = null) =
+    algorithm.randomKey(randomnessSourceOverride)