+++ supreme/src/jvmTest/kotlin/at/asitplus/signum/supreme/symmetric/Test.kt
@@ -67,7 +67,7 @@ val JvmSymmetricTest  by testSuite{
                                 Cipher.getInstance(if (alg.authCapability is AuthCapability.Unauthenticated) "AES/CBC/PKCS5PADDING" else "AES/GCM/NoPadding")
 
                             if (alg is SymmetricEncryptionAlgorithm.AES.GCM) {
-                                val secretKey = alg.randomKey()
+                                val secretKey = alg.randomKey(randomnessSourceOverride = Random.Default)
                                 //GCM need to cast key, because alg is AES with no mode of ops, since we mix CBC and GCM in the test input
                                 val own =
                                     (secretKey as SymmetricKey<AuthCapability.Authenticated<KeyType.Integrated>, NonceTrait.Required, KeyType.Integrated>).andPredefinedNonce(
@@ -105,7 +105,7 @@ val JvmSymmetricTest  by testSuite{
                                     encrypted
                                 )
 
-                                val wrongKey = own.algorithm.randomKey()
+                                val wrongKey = own.algorithm.randomKey(randomnessSourceOverride = Random.Default)
                                 own.decrypt(wrongKey) shouldNot succeed
 
                                 val box = own.algorithm.sealedBox.withNonce(own.algorithm.randomNonce()).from(
@@ -117,7 +117,7 @@ val JvmSymmetricTest  by testSuite{
 
                             } else {
                                 alg as SymmetricEncryptionAlgorithm.AES.CBC.Unauthenticated
-                                val secretKey = alg.randomKey()
+                                val secretKey = alg.randomKey(randomnessSourceOverride = Random.Default)
                                 //CBC
                                 val own = secretKey.encrypt(data).getOrThrow()
                                 jcaCipher.init(
@@ -137,7 +137,7 @@ val JvmSymmetricTest  by testSuite{
                                 own.decrypt(secretKey).getOrThrow() shouldBe jcaCipher.doFinal(encrypted)
 
                                 //this could succeed if we're lucky and padding works out
-                                own.decrypt(own.algorithm.randomKey()).onSuccess {
+                                own.decrypt(own.algorithm.randomKey(randomnessSourceOverride = Random.Default)).onSuccess {
                                     it shouldNotBe data
                                 }
 
@@ -181,7 +181,7 @@ val JvmSymmetricTest  by testSuite{
                         Random.nextBytes(72),
                     ) { data ->
 
-                        val secretKey = alg.randomKey()
+                        val secretKey = alg.randomKey(randomnessSourceOverride = Random.Default)
 
                         //CBC
                         if (alg !is SymmetricEncryptionAlgorithm.AES.WRAP.RFC3394) {
@@ -204,7 +204,7 @@ val JvmSymmetricTest  by testSuite{
                             own.decrypt(secretKey).getOrThrow() shouldBe jcaCipher.doFinal(encrypted)
 
                             //we might get lucky here
-                            own.decrypt(own.algorithm.randomKey()).onSuccess {
+                            own.decrypt(own.algorithm.randomKey(randomnessSourceOverride = Random.Default)).onSuccess {
                                 it shouldNotBe data
                             }
 
@@ -244,7 +244,7 @@ val JvmSymmetricTest  by testSuite{
                                 own.decrypt(secretKey).getOrThrow() shouldBe jcaCipher.doFinal(encrypted)
 
                                 //we might get lucky here
-                                own.decrypt(own.algorithm.randomKey()).onSuccess {
+                                own.decrypt(own.algorithm.randomKey(randomnessSourceOverride = Random.Default)).onSuccess {
                                     it shouldNotBe data
                                 }
 
@@ -273,7 +273,7 @@ val JvmSymmetricTest  by testSuite{
                     Random.nextBytes(32),
                     Random.nextBytes(256),
                 ) { data ->
-                    val secretKey = alg.randomKey()
+                    val secretKey = alg.randomKey(randomnessSourceOverride = Random.Default)
                     val jcaCipher = Cipher.getInstance("ChaCha20-Poly1305");
 
                     val box = if (nonce != null) secretKey.andPredefinedNonce(nonce).getOrThrow().encrypt(data, aad)