+++ CHANGELOG.md
@@ -12,8 +12,7 @@
 * Add COSE_Mac0 support with `CoseMac` class
 * Introduce `ProtectedCoseHeaderSerializer` for serialization/deserialization protected header
 * Replace raw-byte protected header and its ByteArray extension serialization with `CoseHeader` using the new serializer
-* Kotest 6.0.0.M6
-* Remove Kotest Workarounds
+* Allow overriding randomness source for symmetric key generation
 * Extend properties in `JweHeader`
 * Kotlin 2.2.20
 * TestBalloon-powered tests